add delete task

+++ demo/src/main/java/com/example/demo/service/TaskService.java
@@ -1,10 +1,14 @@
 package com.example.demo.service;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.example.demo.dao.TaskDao;
@@ -23,26 +27,43 @@ public class TaskService {
         return taskDao.findByStatus(status);
     }
 
-    public String addTask(Task task) {
+    public ResponseEntity<Map<String, String>> addTask(Task task) {
+        Map<String, String> response = new HashMap<>();
         taskDao.save(task);
-        return "task added";
+        response.put("message", "Task added");
+
+        return ResponseEntity.ok(response);
     }
 
-    public String delete(Task task) {
-        taskDao.delete(task);
-        return "task deleted";
+    public  ResponseEntity<Map<String, String>> delete(int id) {
+        Map<String, String> response = new HashMap<>();
+        Optional<Task> optionalTask = taskDao.findById(id);
+
+        if (optionalTask.isPresent()) {
+            Task task = optionalTask.get();
+            taskDao.delete(task);
+           response.put("message", "Task deleted");
+            return ResponseEntity.ok(response);
+        }else {
+            response.put("message", "Task not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
+        }
     }
 
-    public String updateStatus(int taskID, StatusUpdateDto status) {
+    public ResponseEntity<Map<String, String>> updateStatus(int taskID, StatusUpdateDto status) {
+        Map<String, String> response = new HashMap<>();
         Optional<Task> optionalTask = taskDao.findById(taskID);
+
         if (optionalTask.isPresent()) {
             Task task = optionalTask.get();
             task.setStatus(status.getStatus());
             taskDao.save(task);
-            return "Task status updated";
+            response.put("message", "Task status updated");
+            return ResponseEntity.ok(response);
+        } else {
+            response.put("message", "Task not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
         }
-
-        return "Task not found";
     }
 
 }
